requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
qrcode[pil]>=7.3
Pillow>=8.0
pydantic>=2.0

# AI Agent dependencies
openpyxl>=3.10.0
gitpython>=3.1.27
PyGithub>=1.55
anthropic>=0.25.0
